ATOM/simulations/Feed-in Tariff/requirements.txt
arrow==1.2.3
binaryornot==0.4.4
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cookiecutter==2.1.1
cycler==0.11.0
fonttools==4.38.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
Mesa==1.1.1
networkx==2.8.8
numpy==1.23.5
numpy-financial==1.0.0
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
pyDOE==0.3.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==7.0.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
scipy==1.9.3
six==1.16.0
text-unidecode==1.3
tornado==6.2
tqdm==4.64.1
urllib3==1.26.13
